just started tweet-likes

+++ app.py
@@ -406,7 +406,10 @@ def tweet():
         try:
             conn = mariadb.connect(host=dbcreds.host, password=dbcreds.password, user=dbcreds.user, port=dbcreds.port, database=dbcreds.database)
             cursor = conn.cursor()
-            cursor.execute("DELETE FROM tweet WHERE user_id = '?'",)
+            cursor.execute("SELECT user_id FROM user_session WHERE login_token = ?", [login_token])
+        
+            user_id= cursor.fetchone()[0]
+            cursor.execute("DELETE FROM tweet WHERE id = ? AND user_id = ? ",[tweet_id, user_id])
             conn.commit() 
             rows = cursor.rowcount    
         except Exception as error:
